adds utils function for merge of array
