+++ app/calendars/googlecalendar/googlecalendar.ts
@@ -1,7 +1,7 @@
-import { ExternalEvent } from "app/caldav"
+import { TimeSlotString, mergeArr, convertToExternalEvent ,convertToUnix } from "..utils/"
 import { google } from "googleapis"
 import { createAuthenticatedGoogleOauth } from "./helpers/GoogleClient"
-import { Appointment } from "../appointments/types"
+import { Appointment } from "../../appointments/types"
 import { ConnectedCalendar } from "db"
 import { CalendarService } from "app/calendar-service"
 import { addMilliseconds } from "date-fns"
@@ -36,15 +36,7 @@ export async function createGcalEvent(appointment: Appointment, refreshToken: st
   })
 }
 
-interface TimeSlotString {
-  start: string
-  end: string
-}
 
-interface DateTimeUnix {
-  start: number
-  end: number
-}
 
 export async function getTakenTimeSlots(start: Date, end: Date, refreshToken: string) {
   const calendar = google.calendar({
@@ -79,66 +71,7 @@ export async function getTakenTimeSlots(start: Date, end: Date, refreshToken: st
   return convertToExternalEvent(mergeArr(convertToUnix(rawFreeBusy)))
 }
 
-/**
- * Googlecalendar returns freebusy-slots of multiples calendars. These must be merged.
- * @returns the freebusy slots of all the combined calendars as one
- */
-export function mergeArr(arr: DateTimeUnix[]): DateTimeUnix[] {
-  // Sort the array in descending order
-  arr.sort(function (a, b) {
-    return a.start - b.start
-  })
 
-  let mergedArr: DateTimeUnix[] = []
-
-  // timestamps of the year 2500 to make the algorithm work, see docs for deeper explanation
-  arr.push({ start: 16754814600, end: 16754818200 })
-  let old = arr[0]
-  let temp: DateTimeUnix = { start: -1, end: -1 }
-  for (let el of arr) {
-    let curr = el
-    if (curr === old) continue
-
-    if (curr.start <= old.end) {
-      if (temp.start === -1) {
-        temp.start = old!.start
-      }
-      curr.start = old.start
-    } else {
-      if (temp.start === -1) {
-        temp.start = old.start!
-        temp.end = old.end!
-      } else {
-        temp.end = old.end!
-      }
-      mergedArr.push(temp)
-      temp = { start: -1, end: -1 }
-    }
-
-    old = el
-  }
-  return mergedArr.sort(function (a, b) {
-    return a.start - b.start
-  })
-}
-
-export function convertToUnix(arr: TimeSlotString[]): DateTimeUnix[] {
-  return arr.map((el: TimeSlotString) => {
-    return {
-      start: new Date(el.start).getTime() / 1000,
-      end: new Date(el.end).getTime() / 1000,
-    }
-  })
-}
-
-export function convertToExternalEvent(arr: DateTimeUnix[]): ExternalEvent[] {
-  return arr.map((el: DateTimeUnix) => {
-    return {
-      start: new Date(el.start * 1000),
-      end: new Date(el.end * 1000),
-    }
-  })
-}
 
 export function getCalendarService(calendar: ConnectedCalendar): CalendarService {
   return {
